feat :컨버터 위치 변경 및 수정 요청에 대한 entity convert 메서드 추가, 추가된 isVisible, priority, type 컬럼 반영
Related to : #1
